Extract BinaryExpressionConverter
